requirements.txt
streamlit==1.26.0
langchain==0.0.304
langchain-groq==0.1.2
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
